change allegro notes a bit
I removed the remarks about what OpenGL versions are supported because there is no real notion of not supporting some version of OpenGL in the first place. They are not required to get access to new features. All specifying a versioned context does is cause it to fail if that version is not supported by the system's drivers and hardware.