Added answer for networking direct support question.
Networking is an important part of today's games. Even if just for simple getting and saving data to an external server. There are also quite a few networking libraries on hackage, so finding a suitable one shouldnt be too difficult. Actually adding and implementing the networking library might be hard, but it shouldnt be impossible.